@@ -3,9 +3,10 @@
   "version": "0.1.0",
   "description": "",
   "main": "index.js",
+  "type": "module",
   "scripts": {
-    "dev": "webpack serve --port 3000",
-    "build": "NODE_ENV=production webpack",
+    "dev": "vite serve",
+    "bundle": "vite build --emptyOutDir",
     "lint": "eslint -c .eslintrc.js src/",
     "prettier": "prettier --config .prettierrc.js --write --check .",
     "generate": "graphql-codegen --config codegen.yml"
@@ -14,55 +15,46 @@
   "license": "ISC",
   "devDependencies": {
     "@graphql-codegen/cli": "2.6.2",
-    "@graphql-codegen/typescript": "2.4.10",
-    "@graphql-codegen/typescript-operations": "2.3.7",
-    "@graphql-codegen/typescript-react-apollo": "3.2.13",
-    "@svgr/webpack": "^6.2.1",
-    "@types/lodash": "^4.14.180",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.11",
-    "@typescript-eslint/eslint-plugin": "^5.12.1",
-    "@typescript-eslint/parser": "^5.12.1",
-    "css-loader": "^6.6.0",
-    "eslint": "^8.10.0",
-    "eslint-config-prettier": "^8.4.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.29.2",
-    "eslint-plugin-react-hooks": "^4.3.0",
-    "file-loader": "^6.2.0",
+    "@graphql-codegen/typescript": "2.7.1",
+    "@graphql-codegen/typescript-operations": "2.5.1",
+    "@graphql-codegen/typescript-react-apollo": "3.3.1",
+    "@types/lodash": "^4.14.182",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
+    "@typescript-eslint/eslint-plugin": "^5.30.6",
+    "@typescript-eslint/parser": "^5.30.6",
+    "eslint": "^8.19.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.30.1",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "graphql-tag": "^2.12.6",
-    "html-webpack-plugin": "^5.5.0",
-    "mini-css-extract-plugin": "^2.5.3",
-    "postcss": "^8.4.12",
-    "postcss-loader": "^6.2.1",
+    "postcss": "^8.4.14",
     "postcss-normalize": "^10.0.1",
-    "postcss-preset-env": "^7.4.3",
-    "prettier": "2.5.1",
-    "prettier-plugin-jsdoc": "^0.3.31",
-    "resolve-url-loader": "^5.0.0",
-    "rimraf": "^3.0.0",
-    "sass-loader": "^12.6.0",
-    "ts-loader": "^9.2.6",
-    "typescript": "^4.5.5",
-    "webpack": "^5.69.1",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.7.4"
+    "postcss-preset-env": "^7.7.2",
+    "prettier": "2.7.1",
+    "prettier-plugin-jsdoc": "^0.3.38",
+    "rimraf": "^3.0.2",
+    "typescript": "^4.7.4",
+    "@vitejs/plugin-react": "^2.0.0",
+    "vite-plugin-svgr": "^2.1.0",
+    "vite": "^3.0.0"
   },
   "dependencies": {
-    "@apollo/client": "^3.6.2",
-    "@graphql-codegen/add": "^3.1.1",
-    "@graphql-codegen/cli": "^2.6.2",
-    "@graphql-codegen/fragment-matcher": "^3.2.1",
+    "@apollo/client": "^3.6.9",
+    "@graphql-codegen/add": "^3.2.0",
+    "@graphql-codegen/cli": "2.8.1",
+    "@graphql-codegen/fragment-matcher": "^3.3.0",
     "date-fns": "^2.28.0",
     "debounce": "^1.2.1",
-    "graphql": "^16.4.0",
+    "graphql": "^16.5.0",
     "lodash": "^4.17.21",
     "match-sorter": "^6.3.1",
     "nanoid": "^3.3.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-router-dom": "^6.3.0",
-    "sass": "^1.49.9",
-    "use-debounce": "^7.0.1"
+    "sass": "^1.53.0",
+    "use-debounce": "^8.0.2"
   }
 }
